Should fix all compile warnings in mythtranscode due to replex
git-svn-id: http://svn.mythtv.org/svn/trunk@8103 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/programs/mythtranscode/replex/mpg_common.c
@@ -153,7 +153,7 @@ uint64_t trans_pts_dts(uint8_t *pts)
 }
 
 
-int mring_peek( ringbuffer *rbuf, uint8_t *buf, int l, long off)
+int mring_peek( ringbuffer *rbuf, uint8_t *buf, unsigned int l, uint32_t off)
 {
         int c = 0;
